refactor(config): centralize configuration into a single AppConfig class
