+++ app/components/training_dashboard/logs_panel.py
@@ -2,7 +2,9 @@ import os
 
 import streamlit as st
 
-from config.training_config import TRAINER_STDERR_LOG, TRAINER_STDOUT_LOG
+from config.app_config import get_config
+
+config = get_config()
 
 
 @st.fragment(run_every=1)
@@ -12,13 +14,13 @@ def display_live_logs():
     log_content = ""
     error_count = 0
     try:
-        with open(TRAINER_STDOUT_LOG, "r") as f:
+        with open(config.TRAINER_STDOUT_LOG, "r") as f:
             log_content += f.read()
     except FileNotFoundError:
         log_content += "Waiting for training process to start..."
 
-    if os.path.exists(TRAINER_STDERR_LOG):
-        with open(TRAINER_STDERR_LOG, "r") as f:
+    if os.path.exists(config.TRAINER_STDERR_LOG):
+        with open(config.TRAINER_STDERR_LOG, "r") as f:
             error_content = f.read()
         if error_content:
             log_content += "\n--- ERRORS ---\n" + error_content
