+++ app/view/create_model_view.py
@@ -23,48 +23,55 @@ from backend.manager.global_manager import (
     get_process_manager,
     get_tensorboard_manager,
 )
-from config.training_config import LOCK_FILE
+from config.app_config import get_config
+
+config = get_config()
 
 
 def show_create_model_view():
     """Display the create model interface."""
     # Initialize configuration dictionary
-    config = {}
+    create_model_config = {}
 
     # Get all configuration values
     st.subheader("1. Model Name")
-    config["model_name"] = show_model_name_section()
+    create_model_config["model_name"] = show_model_name_section()
     st.divider()
     st.subheader("2. Fine-tuning Method")
-    config["method"], config["method_params"] = (
+    create_model_config["method"], create_model_config["method_params"] = (
         show_fine_tuning_method_section()
     )
     st.divider()
     st.subheader("3. Data Source")
-    config["data_source"], config["data_config"] = show_data_source_section()
+    create_model_config["data_source"], create_model_config["data_config"] = (
+        show_data_source_section()
+    )
     st.divider()
     st.subheader("4. Model Selection")
-    config["model_config"] = show_model_selection_section()
+    create_model_config["model_config"] = show_model_selection_section()
     st.divider()
     st.subheader("5. Configuration Preview")
-    show_configuration_preview(config)
+    show_configuration_preview(create_model_config)
     st.divider()
     st.subheader("6. Start Training")
-    if show_start_training_section(config):
+    if show_start_training_section(create_model_config):
         # Start the training process
         process_manager = get_process_manager()
         tensorboard_manager = get_tensorboard_manager()
         tensorboard_manager.reset_training_time()
-        process_manager.update_config(config["data_config"], config["model_config"])
+        process_manager.update_config(
+            config["data_config"], config["model_config"]
+        )
         process_manager.start_training()
 
         # Set states to switch to the training dashboard
         st.session_state.session_started_by_app = True
-        st.session_state.model_config = config["model_config"]
+        st.session_state.model_config = create_model_config["model_config"]
         with st.spinner("Waiting for training process to initialize..."):
             start_time = time.time()
             while (
-                not os.path.exists(LOCK_FILE) and time.time() - start_time < 10
+                not os.path.exists(config.LOCK_FILE)
+                and time.time() - start_time < 10
             ):
                 time.sleep(0.5)
